Refactor response handling: split response logic into separate modules
